chore: move getInsatnceXX functions to _instance
