+++ lib/__tests__/_sendEvent.node.test.ts
@@ -1,7 +1,7 @@
 /**
  * @jest-environment node
  */
-import { getInstance } from "../../tests/utils";
+import { getInstance } from "../_instance";
 
 const credentials = {
   apiKey: "testKey",
