+++ lib/browser.ts
@@ -1,10 +1,3 @@
-import AlgoliaAnalytics from "./insights";
-import { processQueue } from "./_processQueue";
+import { getInstanceForBrowser } from "./_instance";
 
-const instance = new AlgoliaAnalytics();
-if (typeof window !== "undefined") {
-  // Process queue upon script execution
-  processQueue.call(instance, window);
-}
-
-export default instance;
+export default getInstanceForBrowser();
