+++ lib/node.ts
@@ -1,5 +1,3 @@
-import AlgoliaAnalytics from "./insights";
-import { getFunctionalInterface } from "./_getFunctionalInterface";
-const instance = new AlgoliaAnalytics();
-const functionalInterface = getFunctionalInterface(instance);
-export default functionalInterface;
+import { getInstanceForNode } from "./_instance";
+
+export default getInstanceForNode();
